+++ kuama_permissions/lib/src/repositories/permissions_preferences_repository.dart
@@ -5,7 +5,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 /// Allows you to read and save user preferences regarding app permissions
 ///
 /// Allows to persist a pair [Permission] - [bool],
-/// [bool] represents whether the [Permission] has already been asked.
+/// [bool] represents whether the [Permission] has already been asked or requested.
 ///
 /// ```dart
 /// final repo = StoragePermissionsRepository();
@@ -18,16 +18,16 @@ import 'package:shared_preferences/shared_preferences.dart';
 class PermissionsPreferencesRepository {
   SharedPreferences get _preferences => GetIt.I();
 
-  /// Returns which permissions have already been asked
-  Map<Permission, bool> checkAsked(List<Permission> permissions) {
+  /// Returns which permissions have already been asked or requested
+  Map<Permission, bool> checkResolved(List<Permission> permissions) {
     return {
       for (final permission in permissions)
         permission: _preferences.getBool('$permission') ?? false,
     };
   }
 
-  /// Update the permissions preferences if they can't be requested
-  Future<void> markAsked(List<Permission> permissions) async {
+  /// Update the permissions preferences if they can't be asked or requested
+  Future<void> markResolved(List<Permission> permissions) async {
     await Future.wait(permissions.map((permission) async {
       await _preferences.setBool('$permission', true);
     }));