+++ kuama_permissions/test/blocs/permissions_state_test.dart
@@ -39,6 +39,7 @@ void main() {
       permissionsStatus: {},
       servicesStatus: {},
       payload: {},
+      isRequested: false,
     );
     final tConfirmableAskState = ConfirmableAskPermissionsState(
       isRefreshing: false,