+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.2.3 (2022-07-19)
+- feat (Permissions) not request permission if it is already requested
+- build update packages version
+
 ## 0.2.0
 
 - feat: Upgrade flutter to 3.0.x and dart to 2.17.x